fix: Passed threads props up to chats page
