+++ components/UserItems.tsx
@@ -13,13 +13,12 @@ export default function UserItems({item, router}: {item: Thread, router: Router}
 
     useEffect(() => {
         getUserInfo();
-    },[item]);
+    },[]);
 
     const openChatRoom = () => {
         const threadID = item.id; 
         router.push({pathname: '/(app)/chatRoom', params: {threadID: threadID, contactName: contact?.firstName, contactID: contact?.id}});
     }
-
 
 
     const getUserInfo = () => {
